Use folly's EliasFano implementation directly.
The timing information is now a bit more involved. Because this "cheats"
and uses the uncompressed postings and encodes them in Elias-Fano on
demand, we have to account for that encoding time. This means that you
have to subtract the two values to get the "real" runtime.

Auto-formatting changed a lot of layout things, and it's too late to go
back and disable them while retaining useful edits.